refactor(cli): normalize guard and hook command layout
