+++ src/envctl/cli/commands/hook/__init__.py
@@ -1,0 +1,1 @@
+"""Hook command package."""
