+++ src/envctl/cli/commands/hook/commands/__init__.py
@@ -1,0 +1,1 @@
+"""Hook subcommands."""
