--- src/envctl/cli/commands/hook_run/__init__.py
@@ -1,1 +1,0 @@
-"""Hook run command package."""
