+++ src/envctl/cli/commands/hooks/commands/__init__.py
@@ -1,0 +1,1 @@
+"""Managed hooks subcommands."""
